requirements.txt
Django==5.2.7
aiogram==3.13.1
python-decouple==3.8